server/requirements.txt
-i https://pypi.org/simple
blinker==1.7.0; python_version >= '3.8'
certifi==2023.11.17; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
flask==3.0.0; python_version >= '3.8'
flask-cors==4.0.0
flask-jwt-extended==4.5.3; python_version >= '3.7' and python_version < '4'
flask-sqlalchemy==3.1.1; python_version >= '3.8'
greenlet==3.0.1; python_version >= '3.7'
gunicorn==21.2.0; python_version >= '3.5'
idna==3.4; python_version >= '3.5'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.3; python_version >= '3.7'
packaging==23.2; python_version >= '3.7'
psycopg2==2.9.9; python_version >= '3.7'
pyjwt==2.8.0; python_version >= '3.7'
python-dotenv==1.0.0; python_version >= '3.8'
requests==2.31.0; python_version >= '3.7'
sqlalchemy==2.0.23; python_version >= '3.7'
typing-extensions==4.8.0; python_version >= '3.8'
urllib3==2.1.0; python_version >= '3.8'
werkzeug==3.0.1; python_version >= '3.8'
